Refactor queue implementation, queue and tech list GUIs
It turns out that iterating existing GUI elements and reordering them is
very fast - only creating or destroying GUI elements is slow. We can use
this to massively simplify the GUI logic and only update each GUI once
per user operation, regardless of the number of techs that were changed.

The queue now only supports adding to the front and back. Adding to the
back is O(1), but adding to the front is O(n).